Changes after comments.

+++ parent/domain/src/main/java/by/issoft/domain/Product.java
@@ -2,6 +2,6 @@ package by.issoft.domain;
 
 public class Product {
     private String name;
-    private Float rate;
-    private Float price;
+    private double rate;
+    private double price;
 }
